Configure the typescript compiler
Note: As of this commit, setting breakpoints and debugging typescript
works in vscode.

Use es5 and commonjs as the 'target' standard and 'module' resolution
systems respectively since that was the default suggestion from:
https://code.visualstudio.com/docs/typescript/typescript-tutorial#_tsconfigjson

Similarly store the content generated by the typescript compiler in
the ./out directory to avoid cluttering the main folder where the
typescript lives by use of 'outDir' in the config.

Also enable 'sourceMap' for debugging purposes per:
https://code.visualstudio.com/docs/typescript/typescript-tutorial#_debugging

Finally ignore the out directory as we don't wish to track potentially
outdated generated content in git.