Enable 'npm test' using jest
Following the advice of:
https://dev.to/s2engineers/testing-with-jest-and-typescript-the-tricky-parts-1gnc

install jest and ts-jest then move the tests into their own
directory (jest seems to think people should have them in __tests__).

Next use the package.json file to configure npm to run tests via the
jest module and configure jest to work with ts-jest and use the
rootDir option to point to where the '__tests__' are.

Finally update the package-lock.json with the newly installed modules.